+++ src/meta/src/rpc/service/stream_service.rs
@@ -138,4 +138,20 @@ where
             table_fragments: info,
         }))
     }
+
+    #[cfg_attr(coverage, no_coverage)]
+    async fn list_tables(
+        &self,
+        request: Request<ListTablesRequest>,
+    ) -> Result<Response<ListTablesResponse>, Status> {
+        let ret = self
+            .catalog_manager
+            .get_tables(&request.into_inner().table_ids)
+            .await;
+        let mut tables = HashMap::default();
+        for table in ret {
+            tables.insert(table.id, table);
+        }
+        Ok(Response::new(ListTablesResponse { tables }))
+    }
 }
