+++ src/storage/compactor/src/compactor_observer/observer_manager.rs
@@ -17,13 +17,13 @@ use std::sync::Arc;
 
 use risingwave_common::system_param::local_manager::LocalSystemParamsManagerRef;
 use risingwave_common_service::observer_manager::{ObserverState, SubscribeCompactor};
-use risingwave_hummock_sdk::filter_key_extractor::{
-    FilterKeyExtractorImpl, FilterKeyExtractorManagerRef,
-};
 use risingwave_pb::catalog::Table;
 use risingwave_pb::meta::relation::RelationInfo;
 use risingwave_pb::meta::subscribe_response::{Info, Operation};
 use risingwave_pb::meta::SubscribeResponse;
+use risingwave_storage::filter_key_extractor::{
+    FilterKeyExtractorImpl, FilterKeyExtractorManagerRef,
+};
 
 pub struct CompactorObserverNode {
     filter_key_extractor_manager: FilterKeyExtractorManagerRef,
