+++ src/storage/src/hummock/compactor/shared_buffer_compact.rs
@@ -22,12 +22,12 @@ use itertools::Itertools;
 use risingwave_common::cache::CachePriority;
 use risingwave_common::catalog::TableId;
 use risingwave_hummock_sdk::compaction_group::StaticCompactionGroupId;
-use risingwave_hummock_sdk::filter_key_extractor::FilterKeyExtractorImpl;
 use risingwave_hummock_sdk::key::{FullKey, UserKey};
 use risingwave_hummock_sdk::key_range::KeyRange;
 use risingwave_hummock_sdk::{CompactionGroupId, HummockEpoch, LocalSstableInfo};
 use risingwave_pb::hummock::compact_task;
 
+use crate::filter_key_extractor::FilterKeyExtractorImpl;
 use crate::hummock::compactor::compaction_filter::DummyCompactionFilter;
 use crate::hummock::compactor::context::CompactorContext;
 use crate::hummock::compactor::{CompactOutput, Compactor};
@@ -100,12 +100,31 @@ pub async fn compact(
 /// For compaction from shared buffer to level 0, this is the only function gets called.
 async fn compact_shared_buffer(
     context: Arc<CompactorContext>,
-    payload: UploadTaskPayload,
+    mut payload: UploadTaskPayload,
 ) -> HummockResult<Vec<LocalSstableInfo>> {
     // Local memory compaction looks at all key ranges.
+
+    let mut existing_table_ids: HashSet<u32> = payload
+        .iter()
+        .map(|imm| imm.table_id.table_id)
+        .dedup()
+        .collect();
+
+    assert!(!existing_table_ids.is_empty());
+
+    let multi_filter_key_extractor = context
+        .filter_key_extractor_manager
+        .acquire(existing_table_ids.clone())
+        .await?;
+    if let FilterKeyExtractorImpl::Multi(multi) = &multi_filter_key_extractor {
+        existing_table_ids = multi.get_exsting_table_ids();
+    }
+    let multi_filter_key_extractor = Arc::new(multi_filter_key_extractor);
+
     let mut size_and_start_user_keys = vec![];
     let mut compact_data_size = 0;
     let mut builder = DeleteRangeAggregatorBuilder::default();
+    payload.retain(|imm| existing_table_ids.contains(&imm.table_id.table_id));
     for imm in &payload {
         let data_size = {
             let tombstones = imm.get_delete_range_tombstones();
@@ -156,20 +175,6 @@ async fn compact_shared_buffer(
             }
         }
     }
-
-    let existing_table_ids: HashSet<u32> = payload
-        .iter()
-        .map(|imm| imm.table_id.table_id)
-        .dedup()
-        .collect();
-
-    assert!(!existing_table_ids.is_empty());
-
-    let multi_filter_key_extractor = context
-        .filter_key_extractor_manager
-        .acquire(existing_table_ids)
-        .await;
-    let multi_filter_key_extractor = Arc::new(multi_filter_key_extractor);
 
     let parallelism = splits.len();
     let mut compact_success = true;
