+++ src/storage/src/lib.rs
@@ -59,6 +59,7 @@ pub mod store_impl;
 pub mod table;
 pub mod write_batch;
 
+pub mod filter_key_extractor;
 pub mod mem_table;
 #[cfg(test)]
 #[cfg(feature = "failpoints")]
